pyproject.toml
[tool.poetry]
name = "cognitive-services-python-sdk-samples"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvd.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-cognitiveservices-knowledge-qnamaker = "0.3.0"
azure-cognitiveservices-language-luis = "^0.7.0"
azure-cognitiveservices-language-spellcheck = "^2.0.0"
azure-ai-textanalytics = "5.0.0"
azure-cognitiveservices-search-autosuggest = "0.1.0"
azure-cognitiveservices-search-customsearch = "^0.3.0"
azure-cognitiveservices-search-entitysearch = "^2.0.0"
azure-cognitiveservices-search-imagesearch = "^2.0.0"
azure-cognitiveservices-search-newssearch = "2.0.0"
azure-cognitiveservices-search-videosearch = "^2.0.0"
azure-cognitiveservices-search-visualsearch = "0.2.0"
azure-cognitiveservices-search-websearch = "^2.0.0"
azure-cognitiveservices-vision-computervision = "0.3.0"
azure-cognitiveservices-vision-contentmoderator = "1.0.0"
azure-cognitiveservices-vision-customvision = "0.4.0"
azure-cognitiveservices-vision-face = "0.6.0"
azure-cognitiveservices-inkrecognizer = "1.0.0b1"
azure-ai-anomalydetector = "3.0.0b2"
azure-ai-vision = "0.15.1b1"
azure-ai-vision-imageanalysis = "1.0.0b1"
pandas = "^2.2.2"
pydantic-settings = "^2.2.1"
aiohttp = "^3.9.5"
aiofiles = "^23.2.1"
loguru = "^0.7.2"
matplotlib = "^3.9.0"
pillow = "^10.3.0"
numpy = "^1.26.4"
azure-ai-language-questionanswering = "^1.1.0"
azure-ai-language-conversations = "^1.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.5"
ipython = "^8.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
ignore-init-module-imports = true

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B", "UP"]
ignore = ["E203", "E266", "E501", "E402", "UP007"]

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["pydantic.validator"]
